requirements.txt
flask==3.1.0
pytest==9.0.2
supabase==2.28.3
bcrypt==5.0.0
python-dotenv==1.2.2
transformers==5.5.1
pillow==12.2.0
PyJWT==2.8.0
torch==2.11.0+cu130
torchvision==0.26.0+cu130